fixed: offline wifi deadlock
Resolves: #2

The program was soft-deadlocking the Mutex when connected to a wifi access
point that couldn't reach the internet. This was caused by acquiring the
Mutex's lock immediately before sending an HTTP request. The request
would hang for a few seconds, and the lock would stay acquired. Since
the `ui` function required the lock in order to work, this caused
awkward hiccups for the end-user.

I fixed this by acquiring the lock only after finishing the request.